add interface to account service
