+++ src/main/java/com/app/atmsimulation/service/AccountService.java
@@ -1,20 +1,12 @@
 package com.app.atmsimulation.service;
 
 import com.app.atmsimulation.model.Account;
-import com.app.atmsimulation.repository.AccountRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 
-public class AccountService {
+public interface AccountService {
 
-    @Autowired
-    private AccountRepository accountRepository;
+    public Account login(String accountNumber, String pin);
 
-    public Account login(String accountNumber, String pin) {
-        
-        return accountRepository.findByAccountNumberAndPin(accountNumber, pin);
-    }
+    public void withdraw();
 
-    public void withdraw() {};
-
-    public void transfer(String destinationAccountNumber) {};
+    public void transfer(String destinationAccountNumber);
 }
