+++ src/main/java/com/app/atmsimulation/service/AccountServiceImpl.java
@@ -1,0 +1,20 @@
+package com.app.atmsimulation.service;
+
+import com.app.atmsimulation.model.Account;
+import com.app.atmsimulation.repository.AccountRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+
+public class AccountServiceImpl implements AccountService {
+
+    @Autowired
+    private AccountRepository accountRepository;
+
+    public Account login(String accountNumber, String pin) {
+        
+        return accountRepository.findByAccountNumberAndPin(accountNumber, pin);
+    }
+
+    public void withdraw() {};
+
+    public void transfer(String destinationAccountNumber) {};
+}
